refactor: API_URL, callback 수정
